refactor(modbus): extract setup-dialog choice value-types into choices.rs
Stage 4 of #89. Moves DialogMode/Transport/Parity/ReconnectChoice/U8Choice and their
ToLabel/config mappings out of setup_dialog.rs. setup_dialog.rs 1089->989.
